convert int32 into int64 when test in windows

+++ cttpunctuator/src/punctuator.py
@@ -72,8 +72,8 @@ class CT_Transformer:
             mini_sentence = cache_sent + mini_sentence
 
             if platform.system() == "Windows":
-                mini_sentence_id = np.array(cache_sent_id + mini_sentence_id, dtype="int64")
-                text_lengths = np.array([len(mini_sentence)], dtype="int64")
+                mini_sentence_id = np.array(cache_sent_id + mini_sentence_id, dtype="int32")
+                text_lengths = np.array([len(mini_sentence)], dtype="int32")
             else:
                 mini_sentence_id = np.array(cache_sent_id + mini_sentence_id, dtype="int64")
                 text_lengths = np.array([len(mini_sentence)], dtype="int32")
